Move adapters to new folder
