+++ app/src/main/java/com/example/scanpal/Fragments/YourEventPageFragment.java
@@ -19,7 +19,7 @@ import androidx.navigation.fragment.NavHostFragment;
 import com.example.scanpal.Models.Event;
 import com.example.scanpal.Controllers.EventController;
 import com.example.scanpal.Callbacks.EventFetchByUserCallback;
-import com.example.scanpal.Controllers.EventGridAdapter;
+import com.example.scanpal.Adapters.EventGridAdapter;
 import com.example.scanpal.Callbacks.EventsFetchCallback;
 import com.example.scanpal.R;
 import com.example.scanpal.Controllers.UserController;
