+++ javadocs/com/example/scanpal/EventGridAdapter.html
@@ -72,7 +72,7 @@ loadScripts(document, 'script');</script>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
 <div class="inheritance">android.widget.BaseAdapter
-<div class="inheritance">com.example.scanpal.Controllers.EventGridAdapter</div>
+<div class="inheritance">com.example.scanpal.Adapters.EventGridAdapter</div>
 </div>
 </div>
 <section class="class-description" id="class-description">
